# backend/src/test/resources/contract/openapi.json
{
  "components" : {
    "schemas" : {
      "AddAddressRequest" : {
        "properties" : {
          "city" : {
            "minLength" : 1,
            "type" : "string"
          },
          "detail" : {
            "minLength" : 1,
            "type" : "string"
          },
          "isDefault" : {
            "type" : "boolean"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "province" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "city", "detail", "name", "phone", "province" ],
        "type" : "object"
      },
      "Address" : {
        "properties" : {
          "city" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "isDefault" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CartItemResponse" : {
        "properties" : {
          "addedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "selected" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "OrderItem" : {
        "properties" : {
          "productId" : {
            "type" : "string"
          },
          "productName" : {
            "type" : "string"
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "unitPrice" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "OrderResponse" : {
        "properties" : {
          "cancelReason" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "estimatedDelivery" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderItem"
            },
            "type" : "array"
          },
          "refundId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "tracking" : {
            "$ref" : "#/components/schemas/OrderTracking"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OrderTracking" : {
        "properties" : {
          "carrier" : {
            "type" : "string"
          },
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/TrackingEvent"
            },
            "type" : "array"
          },
          "trackingNumber" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PageOrderResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageRefundResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/RefundResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "RefundRequest" : {
        "properties" : {
          "amount" : {
            "minimum" : 0.01,
            "type" : "number"
          },
          "reason" : {
            "maxLength" : 200,
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "amount", "reason" ],
        "type" : "object"
      },
      "RefundResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "orderId" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TrackingEvent" : {
        "properties" : {
          "at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateAddressRequest" : {
        "properties" : {
          "city" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "isDefault" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserResponse" : {
        "properties" : {
          "addresses" : {
            "items" : {
              "$ref" : "#/components/schemas/Address"
            },
            "type" : "array"
          },
          "avatarUrl" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "openId" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/refunds" : {
      "get" : {
        "operationId" : "listByStatus",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageRefundResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-refund-controller" ]
      }
    },
...
    "/api/orders" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "userId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}" : {
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/cancel" : {
      "post" : {
        "operationId" : "cancel",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "reason",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/confirm-receive" : {
      "post" : {
        "operationId" : "confirmReceive",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/pay" : {
      "post" : {
        "operationId" : "pay",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/rebuy" : {
      "post" : {
        "operationId" : "rebuy",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CartItemResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/refund" : {
      "post" : {
        "operationId" : "requestRefund",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefundRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RefundResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/remind-ship" : {
      "post" : {
        "operationId" : "remindShip",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/ship" : {
      "post" : {
        "operationId" : "ship",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/api/orders/{id}/tracking" : {
      "get" : {
        "operationId" : "getTracking",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderTracking"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
...
    "/api/users/{userId}/addresses" : {
      "post" : {
        "operationId" : "addAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "user-controller" ]
      }
    },
    "/api/users/{userId}/addresses/{addressId}" : {
      "delete" : {
        "operationId" : "removeAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "user-controller" ]
      },
      "put" : {
        "operationId" : "updateAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "user-controller" ]
      }
    },
    "/api/users/{userId}/addresses/{addressId}/default" : {
      "post" : {
        "operationId" : "setDefault",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "user-controller" ]
      }
    }
  }
}